# docs/swagger/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/rol/delete/{uid}": {
      "delete": {
        "operationId": "RolController_delete",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Rol"
        ]
      }
    },
    "/api/user": {
      "post": {
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDefaultRegisterDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      },
      "get": {
        "operationId": "UserController_findAll",
        "parameters": [
          {
            "name": "orderProperty",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "v_e",
                "ci",
                "first_name",
                "middle_name",
                "first_surname",
                "last_surname",
                "sex",
                "phone",
                "email"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      },
      "patch": {
        "operationId": "UserController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/api/user/protect": {
      "post": {
        "operationId": "UserController_createProtect",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegisterDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
...
    "/api/user/profile": {
      "get": {
        "operationId": "UserController_profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/api/user/profile/data": {
      "patch": {
        "operationId": "UserController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateProfileDataDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/api/user/profile/email": {
      "patch": {
        "operationId": "UserController_updateProfileEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateProfileEmailDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/api/user/profile/password": {
      "patch": {
        "operationId": "UserController_updateProfilePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateProfilePasswordDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/api/user/profile/unregister": {
      "delete": {
        "operationId": "UserController_unregister",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/api/user/delete/{uid}": {
      "delete": {
        "operationId": "UserController_remove",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/api/user/recoveryPassword": {
      "post": {
        "operationId": "UserController_recovery",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRecoveryPasswordDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
...
    "/api/user/newPassword": {
      "post": {
        "operationId": "UserController_newPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserNewPasswordDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/api/user/updatePassword": {
      "patch": {
        "operationId": "UserController_newPasswordUpdate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdatePasswordDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/api/user/activated": {
      "post": {
        "operationId": "UserController_activatedAccount",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserActivateCountDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
...
  },
  "info": {
    "title": "Documentación API",
    "description": "Esta es la documentación de la API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Role": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "typeRol": {
            "enum": [
              "admin",
              "user"
            ],
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "boolean"
          },
          "users": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "uid",
          "name",
          "description",
          "typeRol",
          "permissions",
          "status",
          "users"
        ]
      },
...
      "MedicationRequest": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string"
          },
          "photo_recipe": {
            "type": "string"
          },
          "photo_medical_report": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "delivery_status": {
            "type": "string",
            "enum": [
              "en espera",
              "aceptada",
              "rechazada",
              "entregado"
            ]
          },
          "msg": {
            "type": "string"
          },
          "status": {
            "type": "boolean"
          },
          "uidUser": {
            "type": "string"
          },
          "uidPharmacy": {
            "type": "string"
          },
          "uidPatient": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "pharmacy": {
            "$ref": "#/components/schemas/Pharmacy"
          },
          "patients": {
            "$ref": "#/components/schemas/Patients"
          },
          "deliveryData": {
            "$ref": "#/components/schemas/DeliveryData"
          }
        },
        "required": [
          "uid",
          "photo_recipe",
          "photo_medical_report",
          "code",
          "delivery_status",
          "msg",
          "status",
          "uidUser",
          "uidPharmacy",
          "uidPatient",
          "user",
          "pharmacy",
          "patients",
          "deliveryData"
        ]
      },
      "Pharmacy": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "state": {
            "enum": [
              "Amazonas",
              "Anzoátegui",
              "Apure",
              "Aragua",
              "Barinas",
              "Bolívar",
              "Carabobo",
              "Cojedes",
              "Delta Amacuro",
              "Distrito Capital",
              "Falcón",
              "Guarico",
              "Lara",
              "Mérida",
              "Miranda",
              "Monagas",
              "Nueva Esparta",
              "Portuguesa",
              "Sucre",
              "Táchira",
              "Trujillo",
              "Vargas",
              "Yaracuy",
              "Zulia"
            ],
            "type": "string"
          },
          "municipality": {
            "enum": [
              "Alto Orinoco",
              "Atabapo",
              "Atures",
              "Autana",
              "Manapiare",
              "Maroa",
              "Río Negro",
              "Anaco",
              "Aragua",
              "Carvajal",
              "Diego Bautista Urbaneja",
              "Freites",
              "Guanipa",
              "Guanta",
              "McGregor",
              "Peñalver",
              "San Juan de Capistrano",
              "Santa Ana",
              "Achaguas",
              "Biruaca",
              "Muñoz",
              "Pedro Camejo",
              "San Fernando",
              "Camatagua",
              "Francisco Linares Alcántara",
              "José Ángel Lamas",
              "José Félix Ribas",
              "José Rafael Revenga",
              "Mario Briceño Iragorry",
              "Ocumare de la Costa de Oro",
              "San Casimiro",
              "San Sebastián",
              "Santiago Mariño",
              "Santos Michelena",
              "Alberto Arvelo Torrealba",
              "Antonio José de Sucre",
              "Barinas",
              "Cruz Paredes",
              "Obispos",
              "Pedraza",
              "Rojas",
              "Sosa",
              "Angostura",
              "Caroní",
              "El Callao",
              "Gran Sabana",
              "Heres",
              "Píar",
              "Sifontes",
              "Padre Pedro Chien",
              "Bejuma",
              "Carlos Arvelo",
              "Diego Ibarra",
              "Guacara",
              "Mora",
              "Los Guayos",
              "Montalbán",
              "Naguanagua",
              "Puerto Cabello",
              "San Diego",
              "San Joaquín",
              "Valencia",
              "Anzoátegui",
              "Tinaquillo",
              "Girardot",
              "Lima Blanco",
              "Pao de San Juan Bautista",
              "Ricaurte",
              "Rómulo Gallegos",
              "San Carlos",
              "Tinaco",
              "Antonio Díaz",
              "Casacoima",
              "Pedernales",
              "Tucupita",
              "Buchivacoa",
              "Carirubana",
              "Colina",
              "Dabajuro",
              "Democracia",
              "Falcón",
              "Federación",
              "Jacura",
              "Los Taques",
              "Manaure",
              "Mauroa",
              "Monseñor Iturriza",
              "Palmasola",
              "Petit",
              "Píritu",
              "Silva",
              "Tocópero",
              "Unión",
              "Urumaco",
              "Camaguán",
              "Chaguaramas",
              "El Socorro",
              "Infante",
              "Las Mercedes",
              "Mellado",
              "Monagas",
              "Ortiz",
              "Ribas",
              "Roscio",
              "San Gerónimo de Guayabal",
              "San José de Guaribe",
              "Santa María de Ipire",
              "Zaraza",
              "Crespo",
              "Iribarren",
              "Jiménez",
              "Morán",
              "Palavecino",
              "Simón Planas",
              "Torres",
              "Alberto Adriani",
              "Antonio Pinto Salinas",
              "Aricagua",
              "Arzobispo Chacón",
              "Campo Elías",
              "Caracciolo Parra Olmedo",
              "Cardenal Quintero",
              "Guaraque",
              "Julio César Salas",
              "Justo Briceño",
              "Obispo Ramos de Lora",
              "Padre Noguera",
              "Pueblo Llano",
              "Rangel",
              "Rivas Dávila",
              "Santos Marquina",
              "Tovar",
              "Tulio Febres Cordero",
              "Zea",
              "Acevedo",
              "Baruta",
              "Brión",
              "Buroz",
              "Carrizal",
              "Chacao",
              "Cristóbal Rojas",
              "El Hatillo",
              "Guaicaipuro",
              "Lander",
              "Los Salias",
              "Paz Castillo",
              "Pedro Gual",
              "Plaza",
              "Zamora",
              "Acosta",
              "Aguasay",
              "Caripe",
              "Cedeño",
              "Ezequiel Zamora",
              "Maturín",
              "Piar",
              "Punceres",
              "Santa Bárbara",
              "Sotillo",
              "Uracoa",
              "Antolín del Campo",
              "Díaz",
              "García",
              "Gómez",
              "Maneiro",
              "Marcano",
              "Península de Macanao",
              "Tubores",
              "Villalba",
              "Agua Blanca",
              "Araure",
              "Esteller",
              "Guanare",
              "Guanarito",
              "Monseñor José Vicente de Unda",
              "Ospino",
              "Papelón",
              "San Genaro de Boconoíto",
              "San Rafael de Onoto",
              "Santa Rosalía",
              "Turén",
              "Andrés Eloy Blanco",
              "Andrés Mata",
              "Arismendi",
              "Benítez",
              "Bermúdez",
              "Cajigal",
              "Cruz Salmerón Acosta",
              "Mariño",
              "Mejía",
              "Montes",
              "Ribero",
              "Valdez",
              "Antonio Rómulo Costa",
              "Ayacucho",
              "Cárdenas",
              "Córdoba",
              "Fernández Feo",
              "Francisco de Miranda",
              "García de Hevia",
              "Guásimos",
              "Jáuregui",
              "José María Vargas",
              "Junín",
              "Libertad",
              "Lobatera",
              "Michelena",
              "Panamericano",
              "Pedro María Ureña",
              "Rafael Urdaneta",
              "Samuel Darío Maldonado",
              "San Cristóbal",
              "Seboruco",
              "Simón Rodríguez",
              "Torbes",
              "Uribante",
              "San Judas Tadeo",
              "Andrés Bello",
              "Boconó",
              "Candelaria",
              "Carache",
              "Escuque",
              "José Felipe Márquez Cañizalez",
              "Juan Vicente Campos Elías",
              "La Ceiba",
              "Monte Carmelo",
              "Motatán",
              "Pampán",
              "Pampanito",
              "Rafael Rangel",
              "San Rafael de Carvajal",
              "Trujillo",
              "Urdaneta",
              "Valera",
              "Vargas",
              "Arístides Bastidas",
              "Bolívar",
              "Bruzual",
              "Cocorote",
              "Independencia",
              "José Antonio Páez",
              "La Trinidad",
              "Manuel Monge",
              "Nirgua",
              "Peña",
              "San Felipe",
              "Urachiche",
              "Veroes",
              "Almirante Padilla",
              "Baralt",
              "Cabimas",
              "Catatumbo",
              "Colón",
              "Francisco Javier Pulgar",
              "Páez",
              "Jesús Enrique Lossada",
              "Jesús María Semprún",
              "La Cañada de Urdaneta",
              "Lagunillas",
              "Machiques de Perijá",
              "Mara",
              "Maracaibo",
              "Miranda",
              "Rosario de Perijá",
              "San Francisco",
              "Santa Rita",
              "Simón Bolívar",
              "Sucre",
              "Valmore Rodríguez",
              "Libertador"
            ],
            "type": "string"
          },
          "status": {
            "type": "boolean"
          },
          "uidUser": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "medicationRequest": {
            "$ref": "#/components/schemas/MedicationRequest"
          },
          "deliveryData": {
            "$ref": "#/components/schemas/DeliveryData"
          }
        },
        "required": [
          "uid",
          "name",
          "address",
          "state",
          "municipality",
          "status",
          "uidUser",
          "user",
          "medicationRequest",
          "deliveryData"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string"
          },
          "v_e": {
            "enum": [
              "v",
              "e"
            ],
            "type": "string"
          },
          "ci": {
            "type": "string"
          },
          "first_name": {
            "type": "string"
          },
          "middle_name": {
            "type": "string"
          },
          "first_surname": {
            "type": "string"
          },
          "last_surname": {
            "type": "string"
          },
          "sex": {
            "enum": [
              "f",
              "m"
            ],
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "status": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "activatedAccount": {
            "type": "boolean"
          },
          "attemptCount": {
            "type": "number"
          },
          "dataOfAttempt": {
            "type": "string"
          },
          "uidRol": {
            "type": "string"
          },
          "rol": {
            "$ref": "#/components/schemas/Role"
          },
          "uidPharmacy": {
            "type": "string"
          },
          "pharmacyAS": {
            "$ref": "#/components/schemas/Pharmacy"
          },
          "patients": {
            "$ref": "#/components/schemas/Patients"
          },
          "pharmacy": {
            "$ref": "#/components/schemas/Pharmacy"
          },
          "medicationRequest": {
            "$ref": "#/components/schemas/MedicationRequest"
          },
          "pathology": {
            "$ref": "#/components/schemas/Pathology"
          },
          "medicament": {
            "$ref": "#/components/schemas/Medicament"
          }
        },
        "required": [
          "uid",
          "v_e",
          "ci",
          "first_name",
          "middle_name",
          "first_surname",
          "last_surname",
          "sex",
          "phone",
          "email",
          "password",
          "status",
          "code",
          "activatedAccount",
          "attemptCount",
          "dataOfAttempt",
          "uidRol",
          "rol",
          "uidPharmacy",
          "pharmacyAS",
          "patients",
          "pharmacy",
          "medicationRequest",
          "pathology",
          "medicament"
        ]
      },
...
      "UserDefaultRegisterDTO": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string",
            "format": "uuid"
          },
          "v_e": {
            "type": "string",
            "enum": [
              "v",
              "e"
            ]
          },
          "ci": {
            "type": "string",
            "minLength": 6,
            "maxLength": 8,
            "pattern": "/^\\d+$/"
          },
          "first_name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255
          },
          "middle_name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255
          },
          "first_surname": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255
          },
          "last_surname": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255
          },
          "sex": {
            "type": "string",
            "enum": [
              "f",
              "m"
            ]
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "pattern": "regexPhone"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "uid",
          "v_e",
          "ci",
          "first_name",
          "middle_name",
          "first_surname",
          "last_surname",
          "sex",
          "email",
          "phone",
          "password"
        ]
      },
      "UserRegisterDTO": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string",
            "format": "uuid"
          },
          "v_e": {
            "type": "string",
            "enum": [
              "v",
              "e"
            ]
          },
          "ci": {
            "type": "string",
            "minLength": 6,
            "maxLength": 8,
            "pattern": "/^\\d+$/"
          },
          "first_name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255
          },
          "middle_name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255
          },
          "first_surname": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255
          },
          "last_surname": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255
          },
          "sex": {
            "type": "string",
            "enum": [
              "f",
              "m"
            ]
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "pattern": "regexPhone"
          },
          "password": {
            "type": "string"
          },
          "uidRol": {
            "type": "string",
            "format": "uuid"
          },
          "uidPharmacy": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "uid",
          "v_e",
          "ci",
          "first_name",
          "middle_name",
          "first_surname",
          "last_surname",
          "sex",
          "email",
          "phone",
          "password"
        ]
      },
...
      "UserUpdateDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean"
          }
        },
        "required": [
          "status"
        ]
      },
      "UserUpdateProfileDataDTO": {
        "type": "object",
        "properties": {}
      },
      "UserUpdateProfileEmailDTO": {
        "type": "object",
        "properties": {}
      },
      "UserUpdateProfilePasswordDTO": {
        "type": "object",
        "properties": {
          "olPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "olPassword",
          "newPassword"
        ]
      },
      "UserRecoveryPasswordDTO": {
        "type": "object",
        "properties": {}
      },
      "UserNewPasswordDTO": {
        "type": "object",
        "properties": {
          "newPassword": {
            "type": "string"
          },
          "confirmPassword": {
            "type": "string"
          }
        },
        "required": [
          "newPassword",
          "confirmPassword"
        ]
      },
      "UserUpdatePasswordDTO": {
        "type": "object",
        "properties": {
          "newPassword": {
            "type": "string"
          },
          "confirmPassword": {
            "type": "string"
          },
          "uidUser": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "newPassword",
          "confirmPassword",
          "uidUser"
        ]
      },
      "UserActivateCountDTO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "minLength": 13,
            "maxLength": 16,
            "pattern": "/^\\d+$/"
          }
        },
        "required": [
          "code"
        ]
      },
...
    }
  }
}